Copy the entire Console Log.
The plugin was not waiting for the entire Console log to be written. In
most cases this was ok as the plugin was copying most of the log to be
useful; however, when the log was large the plugin did not copy enough
of it to be useful.

Correct this by starting a new thread to copy the log that can wait for
the existing thread to finish writing to the log.